% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetEdgesComp.R
\name{add_enzymes}
\alias{add_enzymes}
\title{Add enzymes catalyzing each reaction from the Generic genome-scale metabolic model of Homo sapiens}
\usage{
add_enzymes(edges_list, rxn_gene)
}
\arguments{
\item{edges_list}{dataframe with metabolites edges list}

\item{rx_gene}{RXN-GENE dataframe extracted from Homo Sapiens generic GEM}
}
\value{
A dataframe with reactions, metabolites edges, and enzymes
}
\description{
This function add a new column with enzymes matching the correspondent reaction
}
\examples{
\dontrun{
data(rxn_gene)
edges_list_w_enzymes <- add_enzymes(edges_list, rxn_gene)}
}
