% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetEdgesComp.R
\name{edges_list}
\alias{edges_list}
\title{Extract produced and consumed metabolites (met) for each reaction (rxn) of the metabolic model}
\usage{
edges_list(infile)
}
\arguments{
\item{infile}{stoichiometric matrix}
}
\value{
A dataframe with metabolites edges list
}
\description{
This function create the edges list necessary to build the graph of metabolites.
It merge the metabolites produced and consumed by each reaction into a new
dataframe.
}
\examples{
\dontrun{
edges_list <- edges_list(stoich_mat)
}

}
