% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetEdgesComp.R
\name{load_mod}
\alias{load_mod}
\title{Load a metabolic model and extract the stoichiometric matrix}
\usage{
load_mod(infile)
}
\arguments{
\item{infile}{Path to the input SBML file}
}
\value{
A stoichiometric matrix
}
\description{
<<<<<<< HEAD
This function loads a model in SBML format and returns a stoichiometric matrix.
}
\examples{
\dontrun{
stoich_mat <- load_mod(system.file("extdata", "kidney.xml", package = "MetabolitesGraphs"))}
=======
This function loads a model in SBML format and return a stoichiometric matrix.
}
\examples{
\dontrun{
stoich_mat <- load_mod(system.file("extdata", "kidney.xml", package = "Met2Graph"))}
>>>>>>> 1024e35300e30c080ab8806ed0d25fbca2d4c229
}
