% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_expression.R
\name{integrate_expression}
\alias{integrate_expression}
\title{Add expression values to enzymes catalyzing each reaction for each sample}
\usage{
integrate_expression(df, expr_dir, extension, tag, output_dir)
}
\arguments{
\item{df}{dataframe of edges list containing, reactions. metabolites and enzymes. The latter named "Enzyme"}

\item{expr_dir}{path to directory containing samples expression data}

\item{extension}{files extension}

\item{tag}{new filename tags}

\item{output_dir}{path to output directory}
}
\value{
Sample specific dataframes with reactions, metabolites edges, enzymes and expression values
}
\description{
This function creates a new file for each sample containing reactions, connected metabolites, enzymes and expression values
}
\examples{
\dontrun{
expr_dir <- system.file("extdata/expression/", package = "MetabolitesGraphs")
output_dir <- "/path/to/ouput/"
integrate_expression(edges_list_no_rec_mets,expr_dir, "txt", "edges-", output_dir)}
}
